extracted commands to constants
