--- src/index.ts
@@ -1,10 +1,0 @@
-import { CommandsAvailable } from '../src/types';
-
-export const options: CommandsAvailable[] = [
-  { name: 'aboutme', description: '🧒 ~ shows a small biography' },
-  { name: 'projects', description: "📚 ~ show a list of project that I'm been involved" },
-  { name: 'linkedin', description: '🟦 ~ show a link to redirect to my LinkedIn Profile' },
-  { name: 'github', description: '⬛ ~ show a link to redirect to my GitHub Profile' },
-  { name: 'contact', description: '📇 ~ show a link to redirect to my GitHub Profile' },
-  { name: 'cv', description: '📄 ~ dowload my cv' },
-];
